Add simple server that returns a string for any request
Required the httpkit server namespace

Defined a handler function that takes all requests and returns a ring response
object with status okay (200) and a hello world string as the body of the
response.

Defined a server to listen for all http requests and pass them to the handler to
process a response.

Defnined a simple function to call the server with a specific port
number (hacky).